cluster: compute the medoid in float64, so a cluster's name is not per-machine
`d = sqrt(2 - 2*cos)` subtracts two nearly equal numbers whenever two members are
similar, which in a tight cluster is most pairs. At cos 0.99 the result keeps a
hundredth of the input's significant digits, so float32's relative error is
amplified by about 2/(2-2cos) -- 30 to 300x in the clusters this actually hit.

Found by comparing one 871-cluster BioCLIP run against the same run recomputed on
another machine: byte-identical vectors, identical membership for all 26,950
keys, and 5 clusters naming a different medoid -- every one of them tight
(internal cosine 0.976 to 0.997). The disagreements were 4e-5 to 3e-4 relative,
hundreds of times float32 epsilon, so they were not ties that noise flipped.
Neither machine was right: float64 agrees with one on 3 and the other on 2.

That matters because cluster.py's own docstring rests on it -- the medoid is
"stable under re-fitting… what makes a cluster a thing you can point at across
two runs" -- and cluster2 names branches by the medoid of its medoids while
export_seriated names folders by medoid. All of it inherited a name that could
change with the BLAS.

descend.py is fixed by the same change, since it imports cluster_centres; its own
geometry was already float64. cluster2.medoid_of needs nothing: argmax(sub @
centre) has no subtraction and so no cancellation.

Two tests: the medoid is the float64 answer whatever dtype arrives, and -- so the
regression is pinned as real rather than hypothetical -- a fixture chosen by
searching 400 seeds for one where float32 genuinely disagrees.

Existing archives keep the names they were written with; this changes what future
runs produce for those five.

+++ code/cluster/cluster.py
@@ -102,7 +102,18 @@ def cluster_centres(X, labels, probabilities, rows):
     out = []
     for cid in sorted(set(labels) - {-1}):
         idx = np.flatnonzero(labels == cid)
-        members = X[idx]
+        # float64, and not for tidiness. `2 - 2*cos` subtracts two nearly equal
+        # numbers whenever two members are similar, which in a *tight* cluster is
+        # most pairs -- at cos 0.99 the result keeps a hundredth of the input's
+        # significant digits, so float32's relative error is amplified by ~2/(2-2cos).
+        # Measured: one 871-cluster run, recomputed on a second machine from
+        # byte-identical vectors with identical membership, named a different
+        # medoid for 5 clusters -- all of them tight (internal cosine 0.976 to
+        # 0.997). Neither machine was right: float64 agrees with one on 3 of the
+        # five and with the other on 2, so each had float32 wrong somewhere. The
+        # medoid is a cluster's identity across runs, and it cannot depend on
+        # which BLAS summed the row.
+        members = X[idx].astype(np.float64)
         centroid = members.mean(0)
 
         d = np.sqrt(np.maximum(0, 2 - 2 * (members @ members.T)))
